fixes for build
