+++ proxy_agent_shared/src/certificate/mod.rs
@@ -1,0 +1,3 @@
+pub mod certificate_helper;
+#[cfg(windows)]
+pub mod certificate_helper_windows;
